Handle decimal numbers for embedded sensor

+++ home-active-manager-parent/embedded-sensor-module/src/main/java/com/pawbla/project/home/embedded/sensor/handler/Handler.java
@@ -1,11 +1,12 @@
 package com.pawbla.project.home.embedded.sensor.handler;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public interface Handler {
     void handle() throws InterruptedException;
-    int getTemperature();
-    int getHumidity();
+    BigDecimal getTemperature();
+    BigDecimal getHumidity();
     LocalDateTime getLastCorrectReadData();
     boolean isError();
     int getErrorCode();
